requirements.txt
aiohttp==2.3.6
async-timeout==2.0.0
chardet==3.0.4
dpkt==1.9.1
multidict==3.3.2
pyroute2==0.4.21
pytap2==1.4.0
six==1.11.0
websocket-client==0.44.0
yarl==0.15.0
